% Generated by roxygen2 (4.0.2): do not edit by hand
\name{enableTracking}
\alias{enableTracking}
\title{Enable tracking for this session}
\usage{
enableTracking(randomID)
}
\arguments{
\item{randomID}{Should be a random string that is stable for a particular
R user.}
}
\description{
Set the installation ID and enable tracking for this session
}
\details{
This should be called from the user's .Rprofile; ideally the
  randomID value passed in should always be the same for a particular R
  user.

  Normally you should not need to write this function call; use enableScimap()
  to generate the code to insert in .Rprofile.
}

