% Generated by roxygen2 (4.0.2): do not edit by hand
\name{enableScimap}
\alias{disableScimap}
\alias{disableScimapThisSession}
\alias{enableScimap}
\title{Enable or revoke permission for usage tracking}
\usage{
enableScimap()

disableScimap()

disableScimapThisSession()
}
\value{
Returns a string of 25 decimal digits.  This
   is a random but fixed number with no meaning.
   except as a unique identifier.
}
\description{
Enable or revoke permission for usage tracking
}
\details{
Enable or revoke permission for the package to
 send usage statistics to \code{scisoft-net-map.isri.cmu.edu}.

 This is run by hand in interactive mode; there is no
 need to put in a script.  Once you have revoked permission with \code{disableScimap},
 the scimapClient package will not track your usage again
 until you run \code{enableScimap()}

 You can temporarily revoke permission (just until R is stopped and restarted)
 with \code{disableScimapThisSession}.

 (The \code{disableScimap} function saves a file called \code{scimap_permission_denied}
 in your R package directory so that the package will remember your revocation between sessions)
}
\section{Functions}{
\itemize{
\item \code{}: Permanently disable sending of packets

\item \code{}: Disable sending packets just for this session (until R is closed and reopened)
}}
\references{
See your usage and others' at \url{http://scisoft-net-map.isri.cmu.edu}
}
\seealso{
\code{\link{previewPacket}}
}

