% Generated by roxygen2 (4.0.2): do not edit by hand
\name{enableScimap}
\alias{disableScimap}
\alias{disableScimapThisSession}
\alias{enableScimap}
\title{Enable or revoke permission for usage tracking}
\usage{
enableScimap()

disableScimap()

disableScimapThisSession()
}
\description{
Enable or revoke permission for usage tracking
}
\details{
Enable or revoke permission for the package to
 send usage statistics to \code{scisoft-net-map.isri.cmu.edu}.

 This is run by hand in interactive mode; there is no
 need to put in a script.  Once you have revoked permission with \code{disableScimap},
 the scimapClient package will not track your usage again
 until you run \code{enableScimap()}

 You can temporarily revoke permission (just until R is stopped and restarted)
 with \code{disableScimapThisSession}.

 The \code{enableScimap} function saves code in your .Rprofile that
 automatically loads this package each time you run R, and sets a random unique
 ID that anonymizes your data.

 \code{disableScimap} removes that code from your .Rprofile.
}
\section{Functions}{
\itemize{
\item \code{}: Permanently disable sending of packets

\item \code{}: Disable sending packets just for this session (until R is closed and reopened)
}}
\references{
See your usage and others' at \url{http://scisoft-net-map.isri.cmu.edu}
}
\seealso{
\code{\link{previewPacket}}
}

