% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getScimapId}
\alias{getScimapId}
\title{Get scimap unique/anonymous ID for your installation of R}
\usage{
getScimapId()
}
\description{
Return your unique installation ID for scimap usage tracking
}
\details{
The scimapClient package identifies each installation of R with
a unique ID to track usage/installation statistics.
}
\references{
See your usage and others' at \url{http://scisoft-net-map.isri.cmu.edu}
}

