% Generated by roxygen2 (4.0.2): do not edit by hand
\name{isEnabledScimap}
\alias{isEnabledScimap}
\title{Does \code{scimapClient} have permission to track usage?}
\usage{
isEnabledScimap()
}
\value{
boolean
}
\description{
Checks if the package has permission to track usage.
}
\details{
Permission can be granted or revoked with \code{\link{enableScimap}()}
or \code{disableScimap()} from the interactive prompt.

If permission is not given, the \code{scimapRegister()} function
returns silently without doing anything.  If permission is
granted, the package sends usage information, but also
returns silently.

Tracking is automatically disabled if the environment variable R_TESTS is
set (even to ""); this prevents spurious packets from being sent during testing.
It's also disabled while packages are being installed (by checking for the
environment variable R_PACKAGE_NAME).
}
\examples{
## Type this in interactive mode to see whether
## the scimapClient package is enabled to collect usage data.
isEnabledScimap()
# returns TRUE or FALSE
}
\references{
See your usage and others' at \url{http://scisoft-net-map.isri.cmu.edu}
}
\seealso{
\code{\link{previewPacket}}
}

