% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{scimapClient}
\alias{scimapClient}
\alias{scimapClient-package}
\title{Collects and reports usage data on R packages}
\description{
Collects and reports usage data on R packages
}
\details{
Authors of scientific software packages do not always have
good ways of documenting how widely used their work is, and
understanding how it is used in conjunction with other packages.

\code{scimapClient} allows users of these packages to help:
it sends anonymous usage tracking information about the R packages
you use to a server which allows anyone to see how
widely used and installed these packages are.  This is helpful
for demonstrating the usefulness of these packages to
their employers and funding agencies.

This tracking is voluntary and anonymous.  To enable tracking
for all future sessions, type \code{\link{enableScimap}()} from the
interactive prompt; to disable tracking in this and future
sessions, type \code{\link{disableScimap}()}.  If tracking is disabled, this
package will do nothing.
}
\examples{
##  Call this to reenable scimap
\dontrun{enableScimap()}

##  Call this to toggle scimap off if you do not want to send data.
\dontrun{disableScimap()}

## Call this to see what kind of information is sent
previewPacket()
}
\references{
See your usage and others' at \url{http://scisoft-net-map.isri.cmu.edu}
}

